+++ src/rex.mart_actions/static/js/lib/martFromContext.js
@@ -1,0 +1,6 @@
+export default function martFromContext({mart}) {
+  // Currently it relies on the variable name
+  // Probably it would be better if we look through all entities and
+  // find out the one of type 'rexmart_inventory'
+  return mart && mart.id ? mart.id : mart;
+}